// admin/src/api/researches/content-types/researches/researches.ts
// Interface automatically generated by schemas-to-ts

import { Tag } from '../../../tag/content-types/tag/tag';
import { Tag_Plain } from '../../../tag/content-types/tag/tag';
import { AdminPanelRelationPropertyModification } from '../../../../common/schemas-to-ts/AdminPanelRelationPropertyModification';

export interface Researches {
  id: number;
  attributes: {
    createdAt: Date;    updatedAt: Date;    publishedAt?: Date;    title?: string;
    author?: string;
    date?: Date;
    tags?: { data: Tag[] };
    link?: string;
  };
}
export interface Researches_Plain {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  title?: string;
  author?: string;
  date?: Date;
  tags?: Tag_Plain[];
  link?: string;
}

export interface Researches_NoRelations {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  title?: string;
  author?: string;
  date?: Date;
  tags?: number[];
  link?: string;
}

export interface Researches_AdminPanelLifeCycle {
  id: number;
  createdAt: Date;  updatedAt: Date;  publishedAt?: Date;  title?: string;
  author?: string;
  date?: Date;
  tags?: AdminPanelRelationPropertyModification<Tag_Plain>;
  link?: string;
}